+++ R/sits_classification.R
@@ -34,28 +34,24 @@
 #' @param  memsize           Memory available for classification (in GB).
 #' @param  multicores        Number of cores to be used for classification.
 #' @param  output_dir        Output directory
+#' @param  version           Version of classification (in case of multiple tests)
 #' @return A tibble with the predicted labels for each input segment.
 #' @examples
 #' \donttest{
 #' # Retrieve the samples for Mato Grosso
-#' # Install the inSitu library
-#' # devtools::install_github("e-sensing/inSitu")
-#' # library(inSitu)
-#'
-#' samples <- inSitu::br_mt_1_8K_9classes_6bands
 #'
 #' # select the bands "ndvi", "evi"
-#' samples_2bands <- sits_select_bands(samples, ndvi, evi)
+#' samples_ndvi <- sits_select_bands(samples_mt_4bands, ndvi)
 #'
 #' #select a random forest model
 #'
-#' rfor_model <- sits_train(samples_2bands, ml_method = sits_rfor())
+#' rfor_model <- sits_train(samples_ndvi, ml_method = sits_rfor())
 #'
 #' # Retrieve a time series (17 years)
 #' data(point_mt_6bands)
 #'
 #' # select the bands "ndvi" and "evi"
-#' point.tb <- sits_select_bands(point_mt_6bands, ndvi, evi)
+#' point.tb <- sits_select_bands(point_mt_6bands, ndvi)
 #'
 #' # classify the point
 #'
@@ -66,23 +62,13 @@
 #' sits_plot(class.tb)
 #'
 #' # Classify a raster file with 23 instances for one year
-#'
-#' # select the bands "ndvi", "evi" from the "inSitu" package
-#' evi_file <- system.file("extdata/Sinop", "Sinop_evi_2014.tif", package = "inSitu")
-#' ndvi_file <- system.file("extdata/Sinop", "Sinop_ndvi_2014.tif", package = "inSitu")
-#'
-#' files <- c(ndvi_file, evi_file)
-#'
-#' # define the timeline for the files
-#' time_file <- system.file("extdata/Sinop", "timeline_2014.txt", package = "inSitu")
-#' timeline_2013_2014 <- scan(time_file, character())
-#'
+#' files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))
 #' # create a data cube based on the information about the files
-#' sinop <- sits_cube(name = "Sinop", timeline = timeline_2013_2014,
-#'                    bands = c("ndvi", "evi"), files = files)
+#' sinop <- sits_cube(name = "Sinop-crop", timeline = timeline_modis_392,
+#' bands = "ndvi", files = files)
 #'
 #' # classify the raster image
-#' sinop_probs <- sits_classify(sinop, ml_model = rfor_model, memsize = 4, multicores = 2)
+#' sinop_probs <- sits_classify(sinop, ml_model = rfor_model, memsize = 2, multicores = 1)
 #'
 #' # label the classified image
 #' sinop_label <- sits_label_classification(sinop_probs)
@@ -103,7 +89,8 @@ sits_classify <- function(data        = NULL,
                           filter      = NULL,
                           multicores  = 2,
                           memsize     = 4,
-                          output_dir  = "./")
+                          output_dir  = "./",
+                          version     = "v1")
 {
     # check if we are running in Windows
     if (.Platform$OS.type != "unix")
@@ -113,7 +100,7 @@ sits_classify <- function(data        = NULL,
     else
         result <- .sits_classify_cube(cube = data, ml_model = ml_model, interval = interval,
                                       filter = filter, memsize = memsize, multicores = multicores,
-                                      output_dir = output_dir)
+                                      output_dir = output_dir, version = version)
 
     return(result)
 }
@@ -135,59 +122,47 @@ sits_classify <- function(data        = NULL,
 #'                           Elements '0' are excluded from window.
 #' @param  variance          Estimated variance of logit of class_probs (Bayesian smoothing parameter).
 #' @param  output_dir        Output directory where to out the file
+#' @param  version           Version of resulting image (in the case of multiple tests)
 #' @return A tibble with metadata about the output RasterLayer objects.
 #' @examples
 #' \donttest{
 #' # Retrieve the samples for Mato Grosso
-#' # Install the inSitu library
-#' # devtools::install_github("e-sensing/inSitu")
-#' # library(inSitu)
+#' # select the band "ndvi"
 #'
-#' samples <- inSitu::br_mt_1_8K_9classes_6bands
-#'
-#' # select the bands "ndvi", "evi"
-#'
-#' samples_2bands <- sits_select_bands(samples, ndvi, evi)
+#' samples_ndvi <- sits_select_bands(samples_mt_4bands, ndvi)
 #'
 #' #select a random forest model
-#' rfor_model <- sits_train(samples_2bands, ml_method = sits_rfor())
+#' rfor_model <- sits_train(samples_ndvi, ml_method = sits_rfor())
 #'
 #' # Classify a raster file with 23 instances for one year
-#' # select the bands "ndvi", "evi" from the "inSitu" package
-#' evi_file <- system.file("extdata/Sinop", "Sinop_evi_2014.tif", package = "inSitu")
-#' ndvi_file <- system.file("extdata/Sinop", "Sinop_ndvi_2014.tif", package = "inSitu")
-#'
-#' files <- c(ndvi_file, evi_file)
-#'
-#' # define the timeline for the files
-#' time_file <- system.file("extdata/Sinop", "timeline_2014.txt", package = "inSitu")
-#' timeline_2013_2014 <- scan(time_file, character())
+#' files <- c(system.file("extdata/raster/mod13q1/sinop-crop-ndvi.tif", package = "sits"))
 #'
 #' # create a data cube based on the information about the files
-#' raster.tb <- sits_cube(name = "Sinop", timeline = timeline_2013_2014,
-#'                        bands = c("ndvi", "evi"), files = files)
+#' sinop <- sits_cube(name = "Sinop-crop", timeline = timeline_modis_392,
+#' bands = "ndvi", files = files)
 #'
-#' # classify the raster image and generate a probability file
-#' raster_probs.tb <- sits_classify(raster.tb, ml_model = rfor_model, memsize = 4, multicores = 2)
+#' # classify the raster image
+#' sinop_probs <- sits_classify(sinop, ml_model = rfor_model, memsize = 2, multicores = 1)
 #'
-#' # label the probability (no smoothing applied by default)
-#' raster_class.tb <- sits_label_classification(raster_probs.tb)
+#' # label the classified image
+#' sinop_label <- sits_label_classification(sinop_probs)
 #'
 #' # plot the raster image
-#' sits_plot_raster(raster_class.tb, time = 1, title = "Sinop-2013-2014")
+#' sits_plot_raster(sinop_label, time = 1, title = "Sinop-2013-2014")
 #'
 #' # smooth the result with a bayesian filter
-#' raster_class_bayes.tb <- sits_label_classification(raster_probs.tb, smoothing = "bayesian")
+#' sinop_bayes <- sits_label_classification(sinop_probs, smoothing = "bayesian")
 #'
 #' # plot the smoothened image
-#' sits_plot_raster(raster_class_bayes.tb, time = 1, title = "Sinop-smooth")
+#' sits_plot_raster(sinop_bayes, time = 1, title = "Sinop-smooth")
 #' }
 #' @export
 sits_label_classification <- function(cube,
                                       smoothing    = "none",
                                       window       = matrix(1, nrow = 3, ncol = 3, byrow = TRUE),
                                       variance     = 20,
-                                      output_dir   = "./") {
+                                      output_dir   = "./",
+                                      version      = "v1") {
 
     # precondition 1 - check if cube has probability data
     file_name <- .sits_cube_file(cube)
@@ -218,7 +193,8 @@ sits_label_classification <- function(cube,
     values <- matrix(NA, nrow = cube_size, ncol = n_labels)
 
     # create metadata for labelled raster cube
-    cube_labels <-  .sits_cube_labelled(cube, smoothing, output_dir)
+    cube_labels <-  .sits_cube_labelled(cube_probs = cube, smoothing = smoothing, output_dir = output_dir,
+                                        version = version)
     # retrieve the files to be written
     out_files   <-  .sits_cube_files(cube_labels)
 
@@ -328,7 +304,7 @@ sits_label_classification <- function(cube,
 #' @return A vector with the predicted labels.
 .sits_classify_distances <- function(distances_DT, class_info.tb, ml_model, multicores) {
     # define the column names
-    attr_names <- names(environment(ml_model)$train_data_DT)
+    attr_names <- names(.sits_distances(environment(ml_model)$data[1,]))
     ensurer::ensure_that(attr_names, length(.) > 0,
                          err_desc = "sits_classify_distances: training data not saved in the model environment")
     # create a data table to store the distances
@@ -395,10 +371,11 @@ sits_label_classification <- function(cube,
 #' @param  memsize         Memory available for classification (in GB).
 #' @param  multicores      Number of cores to be used for classification.
 #' @param  output_dir      Directory for output file
+#' @param  version         Version of the output (for multiple classifications)
 #' @return A tibble with the metadata for the vector of probabilities for classified RasterLayers.
 #'
 .sits_classify_cube <- function(cube, ml_model, interval, filter,
-                                memsize, multicores, output_dir) {
+                                memsize, multicores, output_dir, version) {
 
     if (.sits_cube_service(cube) == "EOCUBES") {
         res <- .sits_classify_eocubes(cube = cube, ml_model = ml_model, interval = interval,
@@ -432,8 +409,10 @@ sits_label_classification <- function(cube,
                          err_desc = "sits_classify: bands in the training samples are different from bands in the cube")
 
     # classify the data
-    cube_probs <- .sits_classify_multicores(cube, samples, ml_model,
-                                            interval, filter, memsize, multicores, output_dir)
+    cube_probs <- .sits_classify_multicores(cube = cube, samples = samples, ml_model = ml_model,
+                                            interval = interval, filter = filter,
+                                            memsize = memsize, multicores = multicores,
+                                            output_dir = output_dir, version = version)
 
     return(cube_probs)
 }
@@ -463,6 +442,7 @@ sits_label_classification <- function(cube,
 #' @param  memsize         Memory available for classification (in GB).
 #' @param  multicores      Number of cores.
 #' @param  output_dir      Output directory
+#' @param  version         Version of result
 #' @return List of the classified raster layers.
 .sits_classify_multicores <-  function(cube,
                                        samples,
@@ -471,14 +451,16 @@ sits_label_classification <- function(cube,
                                        filter,
                                        memsize,
                                        multicores,
-                                       output_dir) {
+                                       output_dir,
+                                       version) {
 
 
     # get the reference r_obj from the existing cube
     r_obj <- .sits_cube_robj(cube)
 
     # create the medata for the classified cube
-    cube_class <- .sits_cube_classified(cube, samples, interval, output_dir)
+    cube_class <- .sits_cube_classified(cube = cube, samples = samples, interval = interval,
+                                        output_dir = output_dir, version = version)
     # find out how many layers per brick
     n_layers  <- length(sits_labels(samples)$label)
 
@@ -507,7 +489,7 @@ sits_label_classification <- function(cube,
     select.lst <- .sits_timeline_raster_indexes(cube, samples, interval)
 
     # get the attribute names
-    attr_names <- names(environment(ml_model)$train_data_DT)
+    attr_names <- names(.sits_distances(environment(ml_model)$data[1,]))
     ensurer::ensure_that(attr_names, length(.) > 0,
                          err_desc = "sits_classify_distances: training data not saved in the model environment")
     # get initial time for classification
@@ -842,7 +824,7 @@ sits_label_classification <- function(cube,
     select.lst <- .sits_timeline_raster_indexes(cube, samples, interval)
 
     # get the attribute names
-    attr_names <- names(environment(ml_model)$train_data_DT)
+    attr_names <- names(.sits_distances(environment(ml_model)$data[1,]))
     ensurer::ensure_that(attr_names, length(.) > 0,
                          err_desc = "sits_classify_distances: training data not saved in the model environment")
     # get initial time for classification